Example of using a "sentinel error"
A sentinel error is an error that signals that processing can't continue
because of problems with the current state.

Sentinel errors are defined at the package lavel. Generally starts with
Err (EOF is one exception). This should be threated as read-only.
Sentinel errors should be used sparingly and you should not return
explicit error types even when using sentinel errors